Enable more warnings on Windows (#18)
* Turn on W4
* Ignore unknown attributes

+++ midiproc/buffer.cpp
@@ -155,7 +155,7 @@ void DeleteReader(buffer_reader_t *reader)
 //
 int Reader_BytesRead(buffer_reader_t *reader)
 {
-    return reader->pos - reader->buffer->data;
+    return static_cast<int>(reader->pos - reader->buffer->data);
 }
 
 //
