+++ src/hexen/am_map.cpp
@@ -906,7 +906,7 @@ void AM_drawFline(fline_t * fl, int color)
                     d = ay - ax / 2;
                     while (true)
                     {
-                        DOT(x, y, color);
+                        DOT(x, y, static_cast<uint8_t>(color));
                         if (x == fl->b.x)
                             return;
                         if (d >= 0)
@@ -923,7 +923,7 @@ void AM_drawFline(fline_t * fl, int color)
                     d = ax - ay / 2;
                     while (true)
                     {
-                        DOT(x, y, color);
+                        DOT(x, y, static_cast<uint8_t>(color));
                         if (y == fl->b.y)
                             return;
                         if (d >= 0)
